+++ test/reporter-cucumber-html/features/success/specs/success.feature.spec.ts
@@ -46,7 +46,8 @@ test('Scenario with attachments', async ({ scenario }) => {
     'image attachment inline',
     'image attachment as file',
   ]);
-  await expect(scenario.getLogs()).toHaveText(['123 some logs']);
+  await expect(scenario.getLogs()).toContainText('123 some logs');
+  await expect(scenario.getLogs()).toContainText('logs from exec');
 });
 
 test('Scenario with all keywords and success hooks', async ({ scenario }) => {